+++ internal/integration/strict_test.go
@@ -4,7 +4,7 @@ import (
 	"context"
 	"testing"
 
-	"github.com/derision-test/go-mockgen/internal/integration/testdata/mocks"
+	"github.com/derision-test/go-mockgen/v2/internal/integration/testdata/mocks"
 	"github.com/stretchr/testify/assert"
 )