+++ internal/mockgen/generation/generate.go
@@ -11,9 +11,9 @@ import (
 	"strings"
 
 	"github.com/dave/jennifer/jen"
-	"github.com/derision-test/go-mockgen/internal/mockgen/consts"
-	"github.com/derision-test/go-mockgen/internal/mockgen/paths"
-	"github.com/derision-test/go-mockgen/internal/mockgen/types"
+	"github.com/derision-test/go-mockgen/v2/internal/mockgen/consts"
+	"github.com/derision-test/go-mockgen/v2/internal/mockgen/paths"
+	"github.com/derision-test/go-mockgen/v2/internal/mockgen/types"
 )
 
 type Options struct {