+++ CHANGELOG.md
@@ -4,9 +4,14 @@
 
 Nothing yet!
 
+## [v2.0.1] - 2024-03-04
+
+- Changed name of module to `github.com/derision-test/go-mockgen/v2`.
+- Updated module to Go 1.22. [#52](https://github.com/derision-test/go-mockgen/pull/52)
+
 ## [v2.0.0] - 2024-03-04
 
-- Updated module to Go 1.22. [#52](https://github.com/derision-test/go-mockgen/pull/52)
+**PULLED**, as module name was not updated properly.
 
 ## [v1.3.7] - 2022-11-11